pushover_service/requirements.txt
redis
python-pushover
